Attempt to implement the CLI. It doesn't work.
This crashes, every time. I am at a loss; I think it is a Swift bug.

Every time the program runs, it crashes with an EXC_BAD_ACCESS in
JOThreadWordsPuzzle(level:) initializer, the first time it accesses the
level parameter (line 18). This never happened previously, when the
initializer was passed an array directly. The issue seems to be that the
value argument passed by GBCli to _parseOption in main.swift is released
too soon by ARC. When given an argument, I run:

	lines += String((value! as NSString).copy() as NSString)

Line 41 logs the lines just fine, but pausing the debugger and running
`po lines` gives the following:

	size=5
	{
		[0] = class name = __NSAtom
		[1] = class name = __NSAtom
		[2] = class name = __NSAtom
		[3] = class name = __NSAtom
		[4] = class name = __NSAtom
	}

`class name = __NSAtom` is the description of the strings, not the
value. It seems like they aren't being stored correctly by the array, or
aren't being given correctly to Swift from Objective-C.

My best guess is that there is some weirdness involving Swift functions
(and closures, I tried those too with the same result) being called by
Objective-C as a block. ARC may be getting confused, or Swift isn't
converting `id`s to `String`s seamlessly.

I'm confused, and frankly too lazy to bother tracking this down.